+++ Classes/SchemaOrg/Article.php
@@ -12,26 +12,32 @@ namespace Portrino\PxSemantic\SchemaOrg;
  * @author Andre Wuttig<wuttig@portrino.de>
  */
 abstract class Article extends CreativeWork {
+
     /**
      * @var string The actual body of the article.
      */
     private $articleBody;
+
     /**
      * @var string Articles may belong to one or more 'sections' in a magazine or newspaper, such as Sports, Lifestyle, etc.
      */
     private $articleSection;
+
     /**
      * @var int The number of words in the text of the Article.
      */
     private $wordCount;
+
     /**
      * @var string The page on which the work ends; for example "138" or "xvi".
      */
     private $pageEnd;
+
     /**
      * @var string The page on which the work starts; for example "135" or "xiii".
      */
     private $pageStart;
+
     /**
      * @var string Any description of pages that is not separated into pageStart and pageEnd; for example, "1-6, 9, 55" or "10-12, 46-49".
      */