+++ Classes/SchemaOrg/CollectionPage.php
@@ -3,13 +3,13 @@
 namespace Portrino\PxSemantic\SchemaOrg;
 
 /**
- * A WebSite is a set of related web pages and other items typically served from a single web domain and accessible via URLs.
+ * Web page type: Collection page.
  *
- * @see http://schema.org/WebSite Documentation on Schema.org
+ * @see http://schema.org/CollectionPage Documentation on Schema.org
  *
  * @author Andre Wuttig<wuttig@portrino.de>
  */
-class WebSite extends CreativeWork {
+class CollectionPage extends WebPage {
 
     /**
      * @var int