+++ Classes/SchemaOrg/ItemPage.php
@@ -3,13 +3,13 @@
 namespace Portrino\PxSemantic\SchemaOrg;
 
 /**
- * A WebSite is a set of related web pages and other items typically served from a single web domain and accessible via URLs.
+ * A page devoted to a single item, such as a particular product or hotel.
  *
- * @see http://schema.org/WebSite Documentation on Schema.org
+ * @see http://schema.org/ItemPage Documentation on Schema.org
  *
  * @author Andre Wuttig<wuttig@portrino.de>
  */
-class WebSite extends CreativeWork {
+class ItemPage extends WebPage {
 
     /**
      * @var int